+++ fs/src/util/consts.ts
@@ -4,6 +4,8 @@
 export const fileConflictStrategyType = {
   renameWithTilde: 'renameWithTilde',
   renameWithNumber: 'renameWithNumber',
+  renameWithDatetime: 'renameWithDatetime',
+  renameWithEpochMs: 'renameWithEpochMs',
   overwrite: 'overwrite',
   skip: 'skip',
   error: 'error',