+++ fs/src/util/safecopy.ts
@@ -3,22 +3,13 @@ import * as Spec from '$spec'; // Import Spec
 import { FileSpec, FolderSpec, type FSSpec, SymlinkSpec } from '$spec'; // Import Spec
 import type { WalkOptions } from '$walk';
 import { walk } from '$walk'; // Import walk from $walk
+import { DateEx } from '@epdoc/datetime';
 import { _ } from '@epdoc/type';
 import path from 'node:path';
 import type { FilePath, Path } from '../types.ts';
 import { fileConflictStrategyType } from './consts.ts';
 import { resolveType } from './resolve-type.ts';
 import type * as util from './types.ts';
-
-/**
- * Checks if a given value is a valid FileConflictStrategyType.
- *
- * @param value The value to check.
- * @returns True if the value is a valid FileConflictStrategyType, false otherwise.
- */
-export function isFileConflictStrategyType(value: unknown): value is util.FileConflictStrategyType {
-  return typeof value === 'string' && value in Object.keys(fileConflictStrategyType);
-}
 
 /**
  * Safely copies a source file or folder to a destination.
@@ -109,6 +100,34 @@ export async function safeCopyFile(
         } else if (conflictStrategy.errorIfExists) {
           throw new Err.AlreadyExists('File exists', { path: fsDest.path });
         }
+      } else if (conflictStrategy?.type === fileConflictStrategyType.renameWithDatetime) {
+        const sep = conflictStrategy.separator || '-';
+        const prefix = conflictStrategy.prefix || '';
+        const ds = new DateEx().format(conflictStrategy.format || 'yyyyMMddHHmmssSSS');
+        const newPath = new FileSpec(
+          fsDest.dirname,
+          fsDest.basename + sep + prefix + ds + fsDest.extname,
+        );
+        const exists = await newPath.exists();
+        if (exists && conflictStrategy.errorIfExists) {
+          throw new Err.AlreadyExists('File exists', { path: newPath.path });
+        }
+        await fsDest.moveTo(newPath, { overwrite: true });
+        fsDest.clearInfo();
+      } else if (conflictStrategy?.type === fileConflictStrategyType.renameWithEpochMs) {
+        const sep = conflictStrategy.separator || '-';
+        const prefix = conflictStrategy.prefix || '';
+        const ds = new Date().getTime().toString();
+        const newPath = new FileSpec(
+          fsDest.dirname,
+          fsDest.basename + sep + prefix + ds + fsDest.extname,
+        );
+        const exists = await newPath.exists();
+        if (exists && conflictStrategy.errorIfExists) {
+          throw new Err.AlreadyExists('File exists', { path: newPath.path });
+        }
+        await fsDest.moveTo(newPath, { overwrite: true });
+        fsDest.clearInfo();
       }
       // For 'overwrite' strategy, no action is needed here, proceed to copy
     } else {